[Flink] Refactor buffer stream to improve readability.
